requirements.txt
scikit_image==0.14.2
neptune_client==0.4.98
pytorch_memlab==0.0.3
numpy==1.18.1
scipy==1.3.0
torch==1.6.0
tqdm==4.35.0
torchvision==0.7.0
easydict==1.9
matplotlib==3.0.3
GPUtil==1.4.0
imageio==2.9.0
Pillow==7.2.0
PyYAML==5.3.1

